moved more enums out of modules_helper
