+++ lbplanner/classes/enums/MODULE_TYPE.php
@@ -1,0 +1,55 @@
+<?php
+// This file is part of the local_lbplanner.
+//
+// Moodle is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Moodle is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
+/**
+ * enum for module type
+ *
+ * @package local_lbplanner
+ * @subpackage enums
+ * @copyright 2024 NecodeIT
+ * @license http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+
+namespace local_lbplanner\enums;
+
+// TODO: revert to native enums once we migrate to php8.
+
+use local_lbplanner\polyfill\Enum;
+
+/**
+ * Module Types
+ */
+class MODULE_TYPE extends Enum {
+    /**
+     * Grundlegende Kompetenz.
+     */
+    const GK = 0;
+    /**
+     * Erweiterte Kompetenz.
+     */
+    const EK  = 1;
+    /**
+     * Test i.e. exam.
+     */
+    const TEST = 2;
+    /**
+     * TODO: ???
+     */
+    const M  = 3;
+    /**
+     * TODO: ???
+     */
+    const NONE = 4;
+}
